Провел оптимизацию кода
