+++ VKApp/Friends/FriendTableViewCell.swift
