+++ VKApp/Friends/FriendsController.swift
@@ -1,0 +1,67 @@
+//
+//  GroupsController.swift
+//  VKApp
+//
+//  Created by Владислав Лихачев on 29.03.2020.
+//  Copyright © 2020 Vladislav Likhachev. All rights reserved.
+//
+
+import UIKit
+
+class FriendsController: UIViewController {
+    
+    var usersBySections: [(letter: String, users: [User])] = []
+    
+    @IBOutlet weak var sorterControl: SorterBarControl!
+    @IBOutlet weak var tableView: UITableView!
+    
+    
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        tableView.dataSource = self
+        tableView.delegate = self
+        tableView.rowHeight = CGFloat(70)
+        
+        let usersByLetter = Dictionary(grouping: Database.getUsersData(), by: { $0.surname.first! })
+        usersBySections = usersByLetter.map({(letter:String($0.key),users: $0.value)}).sorted(by: {$0.letter < $1.letter})
+        sorterControl.letters = Array(usersByLetter.keys.map({String($0)})).sorted(by: <)
+    }
+    
+    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
+        if segue.identifier == "photoAlbumSegue" {
+            let photoAlbumVC = segue.destination as! PhotoAlbumController
+            if let index = tableView.indexPathForSelectedRow {
+                let user = usersBySections[index.section].users[index.row]
+                photoAlbumVC.photos = user.photos
+                photoAlbumVC.title = "\(user.name) \(user.surname)"
+            }
+        }
+    }
+    
+    @IBAction func sorterBarWasChanged(_ sender: SorterBarControl) {
+        let indexPath = IndexPath(row: 0, section: usersBySections.firstIndex(where: {$0.letter == sender.choosedLetter}) ?? 0)
+        self.tableView.scrollToRow(at: indexPath, at: .middle, animated: true)
+        
+    }
+}
+
+extension FriendsController : UITableViewDataSource, UITableViewDelegate {
+    func numberOfSections(in tableView: UITableView) -> Int { self.usersBySections.count }
+    
+    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {  self.usersBySections[section].users.count }
+    
+    func tableView(_ tableView: UITableView, titleForHeaderInSection section: Int) -> String? {
+        return String(self.usersBySections[section].letter)
+    }
+    
+    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
+        let cell = tableView.dequeueReusableCell(withIdentifier: "userCell", for: indexPath) as! FriendTableViewCell
+        let user = usersBySections[indexPath.section].users[indexPath.row]
+        cell.userLabel.text = "\(user.name) \(user.surname)"
+        cell.userLabel.font = .systemFont(ofSize: CGFloat(16))
+        cell.photoView.imageView.image = user.avatar
+        
+        return cell
+    }
+}
+
