+++ VKApp/Friends/PhotoAlbumController.swift
