+++ VKApp/Friends/PhotoCollectionViewCell.swift
