+++ VKApp/Friends/SorterBarControl.swift
@@ -10,37 +10,39 @@ import UIKit
 
 class SorterBarControl: UIControl {
     var letters = [String]()
-    private var letterButtons = [UIButton]()
-    var stackView = UIStackView()
     var choosedLetter = ""
+    
+    private var letterButtons = [UIButton]()
+    private var stackView = UIStackView()
     
     override func layoutSubviews() {
         setupView()
     }
     func setupView() {
-        isUserInteractionEnabled = true
+
+        letterButtons.removeAll()
         for letter in letters {
             let button = UIButton(type: .system)
             button.setTitle(letter, for: .normal)
             button.setTitleColor(.black, for: .normal)
             button.setTitleColor(.white, for: .selected)
-            button.frame = CGRect(x: 0, y: 0, width: 15, height: 10)
             button.addTarget(self, action: #selector(selectLetter(_:)), for: .touchUpInside)
             self.letterButtons.append(button)
         }
-        
+        stackView.removeFromSuperview()
         stackView = UIStackView(arrangedSubviews: self.letterButtons)
         
         self.addSubview(stackView)
-        stackView.frame = CGRect(x: 0, y: 0, width: 20, height: 15*letters.count)
         stackView.spacing = 1
         stackView.axis = .vertical
         stackView.alignment = .center
         stackView.distribution = .fillEqually
-    //    stackView.translatesAutoresizingMaskIntoConstraints = false
-    //    NSLayoutConstraint.activate([
-     //        stackView.centerYAnchor.constraint(equalTo: centerYAnchor)
-     //    ])
+        stackView.translatesAutoresizingMaskIntoConstraints = false
+        NSLayoutConstraint.activate([
+             stackView.centerYAnchor.constraint(equalTo: centerYAnchor),
+             stackView.topAnchor.constraint(equalTo: topAnchor),
+             stackView.bottomAnchor.constraint(equalTo: bottomAnchor)
+         ])
  
     }
