+++ VKApp/Groups/GroupTableViewCell.swift
